+++ docs/scripts/browstorjs.js
@@ -1,4 +1,4 @@
-// BrowstorJS v1.2.0 @ https://github.com/NullixAT/browstorjs
+// BrowstorJS v1.3.0 @ https://github.com/NullixAT/browstorjs
 /**
  * BrowstorJS
  * Persistent key/value data storage for your Browser and/or PWA, promisified, including file support and service worker support, all with IndexedDB.
@@ -32,10 +32,11 @@ class BrowstorJS {
                 if (!msg || !msg.browstorJsGetFileUrl)
                     return false;
                 // @ts-ignore
+                const urlData = msg.browstorJsGetFileUrl;
                 event.source.postMessage({
                     'browstorJsFileUrl': {
                         'key': msg.browstorJsGetFileUrl.key,
-                        'url': fileUrlPrefix + msg.browstorJsGetFileUrl.key + '/' + msg.browstorJsGetFileUrl.dbName
+                        'url': fileUrlPrefix + urlData.key + '/' + urlData.dbName + (urlData.addAntiCacheParam ? '?c=' + Math.random() : '')
                     }
                 });
                 break;
@@ -47,7 +48,7 @@ class BrowstorJS {
                 }
                 const urlSplit = url.split('/');
                 const key = urlSplit[urlSplit.length - 2];
-                const dbName = urlSplit[urlSplit.length - 1];
+                const dbName = urlSplit[urlSplit.length - 1].split('?')[0];
                 // @ts-ignore
                 event.respondWith(new Promise(async function (resolve) {
                     const value = await (await BrowstorJS.open(dbName)).get(key);
@@ -206,9 +207,13 @@ class BrowstorJS {
     /**
      * Get file url to given storage key
      * @param {string} key
-     * @return {Promise<string>}
+     * @param {boolean} nullIfKeyNotExist Return null when the key not exist in db (default is return the url anyway)
+     * @param {boolean} addAntiCacheParam Add ?c=randomnumber to the generated to prevent browser memory-caching
+     * @return {Promise<string|null>} Null if key does not exist and option is enabled
      */
-    async getUrl(key) {
+    async getUrl(key, nullIfKeyNotExist = false, addAntiCacheParam) {
+        if (nullIfKeyNotExist && !(await this.get(key)))
+            return null;
         const self = this;
         // is undefined in private browsing mode in some browsers or in ancient browsers
         if (typeof navigator.serviceWorker === 'undefined') {
@@ -235,21 +240,21 @@ class BrowstorJS {
             // send message to service worker to request the file url
             navigator.serviceWorker.ready.then(function (serviceWorker) {
                 serviceWorker.active.postMessage({
-                    'browstorJsGetFileUrl': { 'dbName': self.dbName, 'key': key }
+                    'browstorJsGetFileUrl': { 'dbName': self.dbName, 'key': key, 'addAntiCacheParam': addAntiCacheParam }
                 });
             });
         });
     }
     /**
      * Get a data uri that can be used as href or src for images
-     * If value in this key is not a file/blob than this will return a blank 1x1 pixel image data uri
      * @param {string} key
+     * @param {string|null} defaultReturn The default value that is returned in case the key does not exist
      * @return {Promise<string>}
      */
-    async getDataUri(key) {
+    async getDataUri(key, defaultReturn = 'data:image/gif;base64,R0lGODlhAQABAAD/ACwAAAAAAQABAAACADs=') {
         const value = await this.convertValue(await this.get(key), 'blob');
         if (!(value instanceof Blob))
-            return 'data:image/gif;base64,R0lGODlhAQABAAD/ACwAAAAAAQABAAACADs=';
+            return defaultReturn;
         return new Promise(function (resolve) {
             const reader = new FileReader();
             // @ts-ignore
@@ -401,7 +406,10 @@ class BrowstorJS {
      * @private
      */
     async convertValue(value, to) {
-        if (value instanceof Blob && to === 'data') {
+        if (value === null || value === undefined) {
+            return null;
+        }
+        else if (value instanceof Blob && to === 'data') {
             return this.blobToBlobDataObject(value);
         }
         else if (to === 'blob' && typeof value === 'object' && typeof value.type !== 'undefined' && value.type === 'browstorJsBlobData') {