+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.3.0] 2023-01-31
+
+* added `addAntiCacheParam` parameter to `getUrl` to add a random string to generated urls to prevent browser memory-caching
+* added `nullIfKeyNotExist` parameter to `getUrl`
+* added `defaultReturn` parameter to `getDataUri`
+* fixed `convertValue` in case of `null/undefined` returns
+
 ## [1.2.0] 2022-08-02
 
 * added `requestPersistentStorage` and `getStorageSpaceInfo`
@@ -7,7 +14,6 @@
 * added warning when image file url support isn't available because of missing service worker
 * added `getDataUri` function
 * fixed safari support for blob file storage by using ArrayBuffer instead
-
 
 ## [1.0.2] 2022-07-28